refactor: move logger to layers.
