+++ backend/product_service/lambda/product_by_id/product_by_id.ts
@@ -1,5 +1,5 @@
 import { getProductById } from "./dynamo_db";
-import { logger } from "./logger";
+import { logger } from '/opt/nodejs/logger';
 
 const headers = {
   'Content-Type': 'application/json',
