+++ backend/product_service/lambda/product_create/product_create.ts
@@ -1,5 +1,5 @@
 import { createProductWithStock } from "./dynamo_db";
-import { logger } from "./logger";
+import { logger } from '/opt/nodejs/logger';
 
 const headers = {
   'Content-Type': 'application/json',
