+++ backend/product_service/lambda/product_list/product_list.ts
@@ -1,5 +1,5 @@
 import { getAllProducts } from "./dynamo_db";
-import { logger } from "./logger";
+import { logger } from '/opt/nodejs/logger';
 
 const headers = {
   'Content-Type': 'application/json',
