+++ backend/product_service/package.json
@@ -5,7 +5,7 @@
     "product_service": "bin/product_service.js"
   },
   "scripts": {
-    "build": "tsc",
+    "build": "tsc && npm --prefix ./lib/layers/nodejs run build",
     "watch": "tsc -w",
     "test": "jest",
     "cdk": "cdk",
